requirements.txt
anthropic==0.24.0
python-telegram-bot==22.1
python-dotenv==1.0.1
selenium==4.21.0
pytest==8.2.1
